Add hide repository group functionality
Repository groups can be marked as hideable via context menu.
The repositories view filter has an additional option for hiding groups
marked as hideable. If the filter is not activated the sorting places
all hideable repository groups behind all unhideable ones.

This is a proof of concept which was submitted to gerrit and abandoned.

Bug: 545123
Change-Id: I8f782ff681e66b3911991110718ee68745307718
Signed-off-by: Alexander Nittka <alex@nittka.de>